更换静态库并且消除所有警告

+++ Capybara/src/Capybara/ImGui/ImGuiLayer.cpp
@@ -1,7 +1,7 @@
 ﻿#include "precomp.h"
 #include "ImGuiLayer.h"
 
-#include "imgui.h"
+// #include "imgui.h"
 
 #define IMGUI_IMPL_API
 #include "backends/imgui_impl_glfw.h"
@@ -73,7 +73,7 @@ namespace Capybara
     {
         ImGuiIO& io = ImGui::GetIO();
         Application& app = Application::Get();
-        io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+        io.DisplaySize = ImVec2((float)app.GetWindow().GetWidth(), (float)app.GetWindow().GetHeight());
 
         // Rendering
         ImGui::Render();
